fix.버그 수정 및 로직변경
장비 레벨업 업그레이드 비용 버그 수정,
튜토리얼을 완료하지 않고 껐다 켰을 때 오프라인 보상 패널이 뜨면서 아무것도 못 하게 되는 버그 수정

+++ Assets/Scripts/Equipment/EquipMentStatus.cs
@@ -26,7 +26,7 @@ public class EquipmentStatus : MonoBehaviour
     public float goldIncrement;
     public float upgradeGoldIncrement;
     public float maxGoldIncrement;
-    public int upgradeGoldCost;
+    public BigInteger upgradeGoldCost;
     public Text equipLevelText;
     public Text rankLevelText;
     public Button EquipButton;
